made parking spot into modal
